+++ models/targeting/genome_simulation.py
@@ -1,6 +1,6 @@
-from state_machine import Target, Domain, Genome
+from genome_classes import Target, Domain, Genome
 
-# simulation parameters
-dt = 0.001    # seconds
-t0 = 0.0      # seconds
-t1 = 1.0      # seconds
+# simulation parameters (time in seconds)
+dt = 0.001
+t0 = 0.0
+t1 = 3600.0
